pyproject.toml
[project]
name = "gorkpy"
version = "0.1.0"
description = ""
requires-python = ">=3.13,<4.0"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "google-genai[aiohttp] (>=1.29.0,<2.0.0)",
    "pyrate-limiter (>=3.9.0,<4.0.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
